+++ sensing/autoware_cuda_pointcloud_preprocessor/include/autoware/cuda_pointcloud_preprocessor/cuda_outlier_filter/cuda_polar_voxel_outlier_filter_node.hpp
@@ -48,8 +48,7 @@ protected:
 
 private:
   CudaPolarVoxelOutlierFilterParameters filter_params_;
-  // rclcpp always returns integer array as std::vector<int64_t>, so declare the same type here
-  std::vector<int64_t> primary_return_types_;  // Return types considered as primary returns
+  std::vector<int> primary_return_types_;  // Return types considered as primary returns
   std::mutex param_mutex_;
 
   // Diagnostics members
